+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - Government document search filter 
 - Unit tests for government document filter and aggregation
 - Script to delete ESC works that don't correspond with works in the PSQL database
-- Docker Compose file to set up local dev environment easier than before
+- Maintenance script for routine database maintenance processes
 ### Fixed
 - Increased processing speed of govDocES script