+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Docker Compose file to set up local dev environment
 - Updated README steps for setting up local dev environment 
 ### Fixed
-- 
+- Updated unit tests for developmentSetup process
 
 ## 2022-09-07 -- v0.10.6
 ### Added